refactor runtime scene callbacks and split state helpers
