+++ scenes/influence_scene.lua
@@ -13,23 +13,23 @@ function InfluenceScene:exit()
 end
 
 function InfluenceScene:update(dt)
-  self.runtime.update(dt)
+  self.runtime.update_influence(dt)
 end
 
 function InfluenceScene:draw()
-  self.runtime.draw()
+  self.runtime.draw_influence()
 end
 
 function InfluenceScene:keypressed(key)
-  self.runtime.keypressed(key)
+  self.runtime.keypressed_influence(key)
 end
 
 function InfluenceScene:mousepressed(x, y, button)
-  self.runtime.mousepressed(x, y, button)
+  self.runtime.mousepressed_influence(x, y, button)
 end
 
 function InfluenceScene:wheelmoved(dx, dy)
-  self.runtime.wheelmoved(dx, dy)
+  self.runtime.wheelmoved_influence(dx, dy)
 end
 
 function InfluenceScene:resize(w, h)
