+++ terraforming_state.lua
@@ -1,269 +1,32 @@
 local TerraformingState = {}
 TerraformingState.__index = TerraformingState
 
-local STAT_KEYS = { "heat", "air", "water", "soil" }
-local MIN_MAGNETOSPHERE_LEVEL = 1
-local MAX_MAGNETOSPHERE_LEVEL = 4
-local DEFAULT_MAGNETOSPHERE_LEVEL = 2
+local Defaults = require("domain.terraforming_defaults")
+local TerraformingMath = require("domain.terraforming_math")
 
-local DEFAULT_TARGETS = {
-  heat = 0,
-  air = 0,
-  water = 0,
-  soil = 0
-}
+local STAT_KEYS = Defaults.STAT_KEYS
+local MIN_MAGNETOSPHERE_LEVEL = Defaults.MIN_MAGNETOSPHERE_LEVEL
+local MAX_MAGNETOSPHERE_LEVEL = Defaults.MAX_MAGNETOSPHERE_LEVEL
+local DEFAULT_MAGNETOSPHERE_LEVEL = Defaults.DEFAULT_MAGNETOSPHERE_LEVEL
+local DEFAULT_TARGETS = Defaults.DEFAULT_TARGETS
+local DEFAULT_STAT_BOUNDS = Defaults.DEFAULT_STAT_BOUNDS
+local DEFAULT_STARTING_RANGES = Defaults.DEFAULT_STARTING_RANGES
+local DEFAULT_HAZARDS = Defaults.DEFAULT_HAZARDS
+local DEFAULT_INDUSTRY_SLOTS = Defaults.DEFAULT_INDUSTRY_SLOTS
+local COUPLING_EDGE_RULES = Defaults.COUPLING_EDGE_RULES
 
-local DEFAULT_STAT_BOUNDS = {
-  heat = { min = -10, max = 10 },
-  air = { min = -10, max = 0 },
-  water = { min = -10, max = 10 },
-  soil = { min = -10, max = 0 }
-}
-
-local DEFAULT_STARTING_RANGES = {
-  heat = { min = -4, max = 4 },
-  air = { min = -7, max = -1 },
-  water = { min = -4, max = 4 },
-  soil = { min = -7, max = -1 }
-}
-
-local DEFAULT_HAZARDS = {
-  {
-    id = "solar_flare",
-    name = "Solar Flare",
-    category = "Radiation Burst",
-    origin = "space",
-    magnetosphere_blockable = true,
-    deltas = { heat = 2, air = 1 }
-  },
-  {
-    id = "comet_crash_landing",
-    name = "Comet Crash Landing",
-    category = "Orbital Impact",
-    origin = "space",
-    magnetosphere_blockable = true,
-    deltas = { soil = -2, water = 2 }
-  },
-  {
-    id = "micrometeor_swarm",
-    name = "Micrometeor Swarm",
-    category = "Orbital Debris",
-    origin = "space",
-    magnetosphere_blockable = true,
-    deltas = { air = -1, soil = -1 }
-  },
-  {
-    id = "upper_atmos_leak",
-    name = "Upper Atmos Leak",
-    category = "Atmospheric Loss",
-    origin = "atmospheric",
-    magnetosphere_blockable = false,
-    deltas = { air = -2, heat = -1 }
-  },
-  {
-    id = "flash_freeze",
-    name = "Flash Freeze",
-    category = "Climate Shock",
-    origin = "climate",
-    magnetosphere_blockable = false,
-    deltas = { heat = -2, water = -2 }
-  },
-  {
-    id = "dust_storm",
-    name = "Dust Storm",
-    category = "Aeolian Event",
-    origin = "climate",
-    magnetosphere_blockable = false,
-    deltas = { air = -1, soil = -2 }
-  },
-  {
-    id = "acid_rain",
-    name = "Acid Rain",
-    category = "Chemical Event",
-    origin = "chemical",
-    magnetosphere_blockable = false,
-    deltas = { water = 2, soil = -2 }
-  },
-  {
-    id = "dry_front",
-    name = "Dry Front",
-    category = "Hydrologic Shift",
-    origin = "climate",
-    magnetosphere_blockable = false,
-    deltas = { water = -2, soil = -1 }
-  },
-  {
-    id = "volcanic_outgassing",
-    name = "Volcanic Outgassing",
-    category = "Geologic Surge",
-    origin = "geologic",
-    magnetosphere_blockable = false,
-    deltas = { heat = 2, air = -1, soil = 1 }
-  },
-  {
-    id = "spore_bloom",
-    name = "Spore Bloom",
-    category = "Biological Bloom",
-    origin = "biological",
-    magnetosphere_blockable = false,
-    deltas = { soil = 2, air = 1 }
-  }
-}
-
-local DEFAULT_INDUSTRY_SLOTS = 4
-local COUPLING_EDGE_RULES = {
-  { source = "heat", target = "water", factor = 1 },
-  { source = "heat", target = "soil", factor = 1 },
-  { source = "air", target = "heat", factor = 1 },
-  { source = "air", target = "water", factor = 1 },
-  { source = "water", target = "soil", factor = 1 },
-  { source = "water", target = "air", factor = 1 },
-  { source = "soil", target = "air", factor = 1 }
-}
-
-local function shallow_copy(input)
-  local out = {}
-  for k, v in pairs(input or {}) do
-    out[k] = v
-  end
-  return out
-end
-
-local function normalize_hazard_entry(hazard, index)
-  local item = hazard or {}
-  local origin = item.origin
-  local blockable = item.magnetosphere_blockable
-
-  if blockable == nil then
-    blockable = origin == "space"
-  end
-  if origin == nil then
-    origin = blockable and "space" or "planetary"
-  end
-
-  local category = item.category
-  if not category or category == "" then
-    category = origin == "space" and "Orbital Event" or "Planetary Event"
-  end
-
-  return {
-    id = item.id or ("hazard_" .. tostring(index or 0)),
-    name = item.name or ("Hazard " .. tostring(index or 0)),
-    category = category,
-    origin = origin,
-    magnetosphere_blockable = blockable and true or false,
-    deltas = shallow_copy(item.deltas or {})
-  }
-end
-
-local function copy_damage_rules(rules)
-  local out = {}
-  for _, rule in ipairs(rules or {}) do
-    table.insert(out, {
-      stat = rule.stat,
-      min = rule.min,
-      max = rule.max,
-      damage = rule.damage,
-      reason = rule.reason
-    })
-  end
-  return out
-end
-
-local function sign(value)
-  if value > 0 then
-    return 1
-  end
-  if value < 0 then
-    return -1
-  end
-  return 0
-end
-
-local function get_coupling_signal_from_distance(distance)
-  if distance <= 0 then
-    return 3
-  elseif distance <= 1 then
-    return 2
-  elseif distance <= 2 then
-    return 1
-  elseif distance <= 5 then
-    return 0
-  elseif distance >= 10 then
-    return -5
-  end
-
-  local whole = math.floor(distance + 0.0001)
-  return -(whole - 5)
-end
-
-local function edge_key(source_key, target_key)
-  return tostring(source_key) .. "->" .. tostring(target_key)
-end
-
-local function get_edge_delta_from_signal(signal, factor)
-  local strength = math.max(0, math.abs(factor or 1))
-  if strength == 0 then
-    return 0
-  end
-  return signal * strength
-end
-
-local function format_signed(value)
-  if value > 0 then
-    return "+" .. tostring(value)
-  end
-  return tostring(value)
-end
-
-local function clamp(value, min_value, max_value)
-  if value < min_value then
-    return min_value
-  end
-  if value > max_value then
-    return max_value
-  end
-  return value
-end
-
-local function make_scaled_deltas(deltas, scale)
-  local scaled = {}
-  for key, value in pairs(deltas or {}) do
-    scaled[key] = value * scale
-  end
-  return scaled
-end
-
-local function reduce_delta_by_block(delta, block_amount)
-  if delta == 0 then
-    return 0, 0
-  end
-
-  local magnitude = math.abs(delta)
-  local reduced = math.max(0, magnitude - math.max(0, block_amount or 0))
-  local direction = delta > 0 and 1 or -1
-  local effective = direction * reduced
-  local blocked = delta - effective
-  return effective, blocked
-end
-
-local function shuffle_in_place(items)
-  for i = #items, 2, -1 do
-    local j = love.math.random(i)
-    items[i], items[j] = items[j], items[i]
-  end
-end
-
-local function damage_rule_triggers(rule, snapshot)
-  local stat_value = (snapshot and snapshot[rule.stat]) or 0
-  if rule.min and stat_value < rule.min then
-    return false
-  end
-  if rule.max and stat_value > rule.max then
-    return false
-  end
-  return true
-end
+local shallow_copy = TerraformingMath.shallow_copy
+local normalize_hazard_entry = TerraformingMath.normalize_hazard_entry
+local copy_damage_rules = TerraformingMath.copy_damage_rules
+local get_coupling_signal_from_distance = TerraformingMath.get_coupling_signal_from_distance
+local edge_key = TerraformingMath.edge_key
+local get_edge_delta_from_signal = TerraformingMath.get_edge_delta_from_signal
+local format_signed = TerraformingMath.format_signed
+local clamp = TerraformingMath.clamp
+local make_scaled_deltas = TerraformingMath.make_scaled_deltas
+local reduce_delta_by_block = TerraformingMath.reduce_delta_by_block
+local shuffle_in_place = TerraformingMath.shuffle_in_place
+local damage_rule_triggers = TerraformingMath.damage_rule_triggers
 
 function TerraformingState.new(config)
   config = config or {}
